Pointcloud-Models/tarl/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision
torchaudio
matplotlib==3.3.4
numpy==1.20.1
tqdm
pytorch-lightning==2.0.7